many small changes

+++ src/routes/users.js
@@ -50,29 +50,35 @@ defaultUsers.router.post('/login', async(req, res, next) => {
 })
 
 defaultUsers.router.post('/avatars', multipartMiddleware, passport.authenticate('jwt', {session: false}), async(req, res, next) => {
-	const img = req.files.avatar;
-	fs.writeFile(__dirname + '/debug.txt', JSON.stringify(req.files), err => {
-		if (err) throw err;
-	})
-
-	let oldImage = (await User.findById(req.user._id)).avatar;
-	oldImage = path.resolve(__dirname, '../uploads/avatars') + '/' + oldImage.split('/avatars/')[1];
-	fs.unlink(oldImage, err => {
-		if (err) throw err;
-	})
-
-	fs.readFile(img.path, (err, data) => {
-		if (err) next(err);
-		const way = path.resolve(__dirname, '../uploads/avatars') + '/' + img.originalFilename;
-		fs.writeFile(way, data, async(err) => {
-			if (err) next(err);
-			const link = 'http://arusremservis.ru/users/avatars/' + img.originalFilename;
-			await User.findOneAndUpdate({_id: req.user._id}, {$set: {
-				avatar: link
-			}});
-			res.json({success: true, link: link});
+	try {
+		const img = req.files.avatar;
+		fs.writeFile(__dirname + '/debug.txt', JSON.stringify(req.files), err => {
+			if (err) throw err;
 		})
-	})
+
+		let oldImage = (await User.findById(req.user._id)).avatar;
+		if (oldImage != '') {
+			oldImage = path.resolve(__dirname, '../uploads/avatars') + '/' + oldImage.split('/avatars/')[1];
+			fs.unlink(oldImage, err => {
+				if (err) throw err;
+			})
+		}
+
+		fs.readFile(img.path, (err, data) => {
+			if (err) next(err);
+			const way = path.resolve(__dirname, '../uploads/avatars') + '/' + img.originalFilename;
+			fs.writeFile(way, data, async(err) => {
+				if (err) next(err);
+				const link = 'http://arusremservis.ru/users/avatars/' + img.originalFilename;
+				await User.findOneAndUpdate({_id: req.user._id}, {$set: {
+					avatar: link
+				}});
+				res.json({success: true, link: link});
+			})
+		})
+	} catch(err) {
+		next(err);
+	}
 })
 
 defaultUsers.router.get('/avatars/:img', async(req, res) => {
